+++ packages/spry/lib/src/middleware/middleware_stack.dart
@@ -1,3 +1,6 @@
+import 'dart:async';
+import 'dart:io';
+
 import '../application.dart';
 import 'middleware.dart';
 
@@ -16,7 +19,7 @@ class MiddlewareStack extends Iterable<Middleware> {
   ///
   /// If [prepend] is `true`, the [middleware] will be added to the beginning of
   /// the stack.
-  void use(Middleware middleware, {prepend = false}) {
+  void use(Middleware middleware, {bool prepend = false}) {
     if (prepend) {
       _application.logger.config('Prepending middleware: $middleware');
 
@@ -26,4 +29,22 @@ class MiddlewareStack extends Iterable<Middleware> {
     _application.logger.config('Appending middleware: $middleware');
     _stack.add(middleware);
   }
+
+  /// Adds a closure style [Middleware] to the stack.
+  void closure(
+    FutureOr<void> Function(HttpRequest request, Next next) process, {
+    bool prepend = false,
+  }) {
+    return use(_ClosureMifddleware(process), prepend: prepend);
+  }
+}
+
+class _ClosureMifddleware implements Middleware {
+  final FutureOr<void> Function(HttpRequest request, Next next) closure;
+
+  const _ClosureMifddleware(this.closure);
+
+  @override
+  Future<void> process(HttpRequest request, Next next) async =>
+      closure(request, next);
 }